refactor: rename localizations directory
