+++ app/localizations/french.ts
