+++ app/localizations/greek.ts
