+++ app/localizations/hungarian.ts
