+++ app/localizations/norwegian.ts
