+++ app/localizations/thai.ts
